+++ knightbus/src/KnightBus.Core/DistributedTracing/IDistributedTracingWriter.cs
@@ -1,7 +1,0 @@
-namespace KnightBus.Core.DistributedTracing;
-
-// Obsolete?
-public interface IDistributedTracingWriter
-{
-    void Set(string traceId);
-}