+++ knightbus/src/KnightBus.Core/IMessagePropertiesProvider.cs
@@ -1,7 +1,0 @@
-namespace KnightBus.Core;
-
-public interface IMessagePropertiesProvider
-{
-    string Get(string key);
-    void Set(string key, string value);
-}